# api/swagger.yaml
openapi: 3.0.1
info:
  title: Emerald Web Api
  version: v1.0.0
servers:
- url: /
security:
- Bearer: []
paths:
...
  /api/Application/GetAccountScoreData:
    post:
      tags:
      - Application
      operationId: apiApplicationGetAccountScoreDataPOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/AccountScoreRequest'
            example:
              CompanyID: 1134
              ApplicationID: 1032
              IsExisintg: false
          application/json:
            schema:
              $ref: '#/components/schemas/AccountScoreRequest'
            example:
              CompanyID: 1134
              ApplicationID: 1032
              IsExisintg: false
          text/json:
            schema:
              $ref: '#/components/schemas/AccountScoreRequest'
            example:
              CompanyID: 1134
              ApplicationID: 1032
              IsExisintg: false
          application/*+json:
            schema:
              $ref: '#/components/schemas/AccountScoreRequest'
            example:
              CompanyID: 1134
              ApplicationID: 1032
              IsExisintg: false
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/GetAccountScoreReport/{companyId}:
    get:
      tags:
      - Application
      operationId: apiApplicationGetAccountScoreReportCompanyIdGET
      parameters:
      - name: companyId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: double
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/CompanyApplications/{companyId}:
    get:
      tags:
      - Application
      operationId: apiApplicationCompanyApplicationsCompanyIdGET
      parameters:
      - name: companyId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: double
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/Create:
    post:
      tags:
      - Application
      operationId: apiApplicationCreatePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/ApplicationCreateRequest'
            example:
              LoanApplied: 15000
              LoanTerm: 12
              LoanPurpose: Stock
              WhenYouNeedLoan: ASAP
              ApplicationStatusId: 3
              HaveBusinessRevenue: true
              MonthlyTotalSales: 2500000
              MainGuarantor:
                GuarantorID: 0
                CompanyID: 0
                ForName: adeel
                SurName: hassan
                Birth: 11-Feb-1994
                Title: Mr
                Mobile: fdsg42135
                Email: adeelhassan592@gmail.com
                Is25PercentShareHolder: false
                Director: false
              CompanyDatas:
                CompanyID: 0
                CompanyName: BANK  LTD
                CoRegNo: "04348058"
                Website: www.BankLTD.com
                BusinessSector: Stock investment
                Email: bankltd@gmail.com
                BusinessType: Ltd Company
                IncorporationDate: 07-Jan-2002
                NoOfDirectors: 13
                NoOf25PercentageShareHolders: 12
              Addressview:
                AddressID: 0
                GuarantorID: 0
                CompanyID: 0
                Property: Bank Gallery
                Street: 13 High Street
                Locality: Kenilworth
                County: Warwickshire
                PostCode: CV8 1LY
                AddressType: 1
                AddressOrder: 0
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationCreateRequest'
            example:
              LoanApplied: 15000
              LoanTerm: 12
              LoanPurpose: Stock
              WhenYouNeedLoan: ASAP
              ApplicationStatusId: 3
              HaveBusinessRevenue: true
              MonthlyTotalSales: 2500000
              MainGuarantor:
                GuarantorID: 0
                CompanyID: 0
                ForName: adeel
                SurName: hassan
                Birth: 11-Feb-1994
                Title: Mr
                Mobile: fdsg42135
                Email: adeelhassan592@gmail.com
                Is25PercentShareHolder: false
                Director: false
              CompanyDatas:
                CompanyID: 0
                CompanyName: BANK  LTD
                CoRegNo: "04348058"
                Website: www.BankLTD.com
                BusinessSector: Stock investment
                Email: bankltd@gmail.com
                BusinessType: Ltd Company
                IncorporationDate: 07-Jan-2002
                NoOfDirectors: 13
                NoOf25PercentageShareHolders: 12
              Addressview:
                AddressID: 0
                GuarantorID: 0
                CompanyID: 0
                Property: Bank Gallery
                Street: 13 High Street
                Locality: Kenilworth
                County: Warwickshire
                PostCode: CV8 1LY
                AddressType: 1
                AddressOrder: 0
          text/json:
            schema:
              $ref: '#/components/schemas/ApplicationCreateRequest'
            example:
              LoanApplied: 15000
              LoanTerm: 12
              LoanPurpose: Stock
              WhenYouNeedLoan: ASAP
              ApplicationStatusId: 3
              HaveBusinessRevenue: true
              MonthlyTotalSales: 2500000
              MainGuarantor:
                GuarantorID: 0
                CompanyID: 0
                ForName: adeel
                SurName: hassan
                Birth: 11-Feb-1994
                Title: Mr
                Mobile: fdsg42135
                Email: adeelhassan592@gmail.com
                Is25PercentShareHolder: false
                Director: false
              CompanyDatas:
                CompanyID: 0
                CompanyName: BANK  LTD
                CoRegNo: "04348058"
                Website: www.BankLTD.com
                BusinessSector: Stock investment
                Email: bankltd@gmail.com
                BusinessType: Ltd Company
                IncorporationDate: 07-Jan-2002
                NoOfDirectors: 13
                NoOf25PercentageShareHolders: 12
              Addressview:
                AddressID: 0
                GuarantorID: 0
                CompanyID: 0
                Property: Bank Gallery
                Street: 13 High Street
                Locality: Kenilworth
                County: Warwickshire
                PostCode: CV8 1LY
                AddressType: 1
                AddressOrder: 0
          application/*+json:
            schema:
              $ref: '#/components/schemas/ApplicationCreateRequest'
            example:
              LoanApplied: 15000
              LoanTerm: 12
              LoanPurpose: Stock
              WhenYouNeedLoan: ASAP
              ApplicationStatusId: 3
              HaveBusinessRevenue: true
              MonthlyTotalSales: 2500000
              MainGuarantor:
                GuarantorID: 0
                CompanyID: 0
                ForName: adeel
                SurName: hassan
                Birth: 11-Feb-1994
                Title: Mr
                Mobile: fdsg42135
                Email: adeelhassan592@gmail.com
                Is25PercentShareHolder: false
                Director: false
              CompanyDatas:
                CompanyID: 0
                CompanyName: BANK  LTD
                CoRegNo: "04348058"
                Website: www.BankLTD.com
                BusinessSector: Stock investment
                Email: bankltd@gmail.com
                BusinessType: Ltd Company
                IncorporationDate: 07-Jan-2002
                NoOfDirectors: 13
                NoOf25PercentageShareHolders: 12
              Addressview:
                AddressID: 0
                GuarantorID: 0
                CompanyID: 0
                Property: Bank Gallery
                Street: 13 High Street
                Locality: Kenilworth
                County: Warwickshire
                PostCode: CV8 1LY
                AddressType: 1
                AddressOrder: 0
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/{id}:
    get:
      tags:
      - Application
      operationId: apiApplicationIdGET
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
...
  /api/Application/Duplicates/{id}:
    get:
      tags:
      - Application
      operationId: apiApplicationDuplicatesIdGET
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int32
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/DuplicateApplicationDetails:
    post:
      tags:
      - Application
      operationId: apiApplicationDuplicateApplicationDetailsPOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/DuplicateLookupRequest'
            example:
              ApplicationID: 1141
              DuplicateApplicationID: 6684
          application/json:
            schema:
              $ref: '#/components/schemas/DuplicateLookupRequest'
            example:
              ApplicationID: 1141
              DuplicateApplicationID: 6684
          text/json:
            schema:
              $ref: '#/components/schemas/DuplicateLookupRequest'
            example:
              ApplicationID: 1141
              DuplicateApplicationID: 6684
          application/*+json:
            schema:
              $ref: '#/components/schemas/DuplicateLookupRequest'
            example:
              ApplicationID: 1141
              DuplicateApplicationID: 6684
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/GetEquifaxData:
    post:
      tags:
      - Application
      operationId: apiApplicationGetEquifaxDataPOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/EquifaxRequest'
            example:
              CompanyID: 3
              IsExisting: true
              IsReport: false
          application/json:
            schema:
              $ref: '#/components/schemas/EquifaxRequest'
            example:
              CompanyID: 3
              IsExisting: true
              IsReport: false
          text/json:
            schema:
              $ref: '#/components/schemas/EquifaxRequest'
            example:
              CompanyID: 3
              IsExisting: true
              IsReport: false
          application/*+json:
            schema:
              $ref: '#/components/schemas/EquifaxRequest'
            example:
              CompanyID: 3
              IsExisting: true
              IsReport: false
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/ApplicationJourney/{id}:
    get:
      tags:
      - Application
      operationId: apiApplicationApplicationJourneyIdGET
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: double
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/ApplicationLoan/{id}:
    get:
      tags:
      - Application
      operationId: apiApplicationApplicationLoanIdGET
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: double
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/addresslookup:
    post:
      tags:
      - Application
      operationId: apiApplicationAddresslookupPOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/AddressLookupRequest'
            example:
              PostCode: SL1 3DG
          application/json:
            schema:
              $ref: '#/components/schemas/AddressLookupRequest'
            example:
              PostCode: SL1 3DG
          text/json:
            schema:
              $ref: '#/components/schemas/AddressLookupRequest'
            example:
              PostCode: SL1 3DG
          application/*+json:
            schema:
              $ref: '#/components/schemas/AddressLookupRequest'
            example:
              PostCode: SL1 3DG
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/companylookup:
    post:
      tags:
      - Application
      operationId: apiApplicationCompanylookupPOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/CompanyLookupRequest'
            example:
              Name: bank
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyLookupRequest'
            example:
              Name: bank
          text/json:
            schema:
              $ref: '#/components/schemas/CompanyLookupRequest'
            example:
              Name: bank
          application/*+json:
            schema:
              $ref: '#/components/schemas/CompanyLookupRequest'
            example:
              Name: bank
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/applicationNote:
    post:
      tags:
      - Application
      operationId: apiApplicationApplicationNotePOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/ApplicationNotesDetailsRequest'
            example:
              ApplicationId: 16992
              CategoryId: 9
              NoteDetails: Lorem Ipsum is simply dummy text of the printing and typesetting
                industry. Lorem Ipsum has been the industry's standard dummy text
                ever since the 1500s, when an unknown printer took a galley of type
                and scrambled it to make a type specimen book. It has survived not
                only five centuries, but also the leap into electronic typesetting,
                remaining essentially unchanged. It was popularised in the 1960s with
                the release of Letraset sheets containing Lorem Ipsum passages, and
                more recently with desktop publishing software like Aldus PageMaker
                including versions of Lorem Ipsum.
              CreatedBy: Test
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationNotesDetailsRequest'
            example:
              ApplicationId: 16992
              CategoryId: 9
              NoteDetails: Lorem Ipsum is simply dummy text of the printing and typesetting
                industry. Lorem Ipsum has been the industry's standard dummy text
                ever since the 1500s, when an unknown printer took a galley of type
                and scrambled it to make a type specimen book. It has survived not
                only five centuries, but also the leap into electronic typesetting,
                remaining essentially unchanged. It was popularised in the 1960s with
                the release of Letraset sheets containing Lorem Ipsum passages, and
                more recently with desktop publishing software like Aldus PageMaker
                including versions of Lorem Ipsum.
              CreatedBy: Test
          text/json:
            schema:
              $ref: '#/components/schemas/ApplicationNotesDetailsRequest'
            example:
              ApplicationId: 16992
              CategoryId: 9
              NoteDetails: Lorem Ipsum is simply dummy text of the printing and typesetting
                industry. Lorem Ipsum has been the industry's standard dummy text
                ever since the 1500s, when an unknown printer took a galley of type
                and scrambled it to make a type specimen book. It has survived not
                only five centuries, but also the leap into electronic typesetting,
                remaining essentially unchanged. It was popularised in the 1960s with
                the release of Letraset sheets containing Lorem Ipsum passages, and
                more recently with desktop publishing software like Aldus PageMaker
                including versions of Lorem Ipsum.
              CreatedBy: Test
          application/*+json:
            schema:
              $ref: '#/components/schemas/ApplicationNotesDetailsRequest'
            example:
              ApplicationId: 16992
              CategoryId: 9
              NoteDetails: Lorem Ipsum is simply dummy text of the printing and typesetting
                industry. Lorem Ipsum has been the industry's standard dummy text
                ever since the 1500s, when an unknown printer took a galley of type
                and scrambled it to make a type specimen book. It has survived not
                only five centuries, but also the leap into electronic typesetting,
                remaining essentially unchanged. It was popularised in the 1960s with
                the release of Letraset sheets containing Lorem Ipsum passages, and
                more recently with desktop publishing software like Aldus PageMaker
                including versions of Lorem Ipsum.
              CreatedBy: Test
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/ApplicationNotes/{applicationId}:
    get:
      tags:
      - Application
      operationId: apiApplicationApplicationNotesApplicationIdGET
      parameters:
      - name: applicationId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: double
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
  /api/Application/search:
    post:
      tags:
      - Application
      operationId: apiApplicationSearchPOST
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/ApplicationsSearchRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationsSearchRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/ApplicationsSearchRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ApplicationsSearchRequest'
      responses:
        200:
          description: Success
      x-swagger-router-controller: Application
...
components:
  schemas:
...
    AccountScoreRequest:
      type: object
      properties:
        companyID:
          type: number
          format: double
        applicationID:
          type: number
          format: double
        isExisintg:
          type: boolean
      additionalProperties: false
    MainGuarantorData:
      type: object
      properties:
        guarantorID:
          type: integer
          format: int32
        companyID:
          type: integer
          format: int32
        forName:
          type: string
        surName:
          type: string
        birth:
          type: string
        status:
          type: string
        title:
          type: string
        mobile:
          type: string
        landline:
          type: string
        email:
          type: string
        shareholder:
          type: string
        is25PercentShareHolder:
          type: boolean
        director:
          type: boolean
      additionalProperties: false
    CompanyData:
      type: object
      properties:
        companyID:
          type: integer
          format: int32
        companyName:
          type: string
        coRegNo:
          type: string
        website:
          type: string
        tradingAs:
          type: string
        businessSector:
          type: string
        email:
          type: string
        businessType:
          type: string
        incorporationDate:
          type: string
        shareholders:
          type: string
        director:
          type: string
        noOfDirectors:
          type: integer
          format: int32
        haveBusinessLease:
          type: boolean
          nullable: true
        instagram:
          type: string
        facebook:
          type: string
        noOf25PercentageShareHolders:
          type: integer
          format: int32
      additionalProperties: false
    AddressData:
      type: object
      properties:
        addressID:
          type: integer
          format: int32
        guarantorID:
          type: integer
          format: int32
        companyID:
          type: integer
          format: int32
        property:
          type: string
        street:
          type: string
        locality:
          type: string
        county:
          type: string
        town:
          type: string
        postCode:
          type: string
        email:
          type: string
        telephone:
          type: string
        businessCategory:
          type: string
        businessSector:
          type: string
        websites:
          type: string
        vatNo:
          type: string
        addressType:
          type: integer
          format: int32
        addressOrder:
          type: integer
          format: int32
        noOfEmployees:
          type: string
      additionalProperties: false
    ApplicationCreateRequest:
      type: object
      properties:
        loanApplied:
          type: number
          format: double
        loanTerm:
          type: integer
          format: int32
        loanPurpose:
          type: string
        whenYouNeedLoan:
          type: string
        applicationStatusId:
          type: number
          format: double
        haveBusinessRevenue:
          type: boolean
          nullable: true
        applyOnBehalf:
          type: boolean
          nullable: true
        haveExistingLoans:
          type: boolean
          nullable: true
        applicationScoreFromWeb:
          type: number
          format: double
          nullable: true
        monthlyTotalSales:
          type: number
          format: double
        mainGuarantor:
          $ref: '#/components/schemas/MainGuarantorData'
        companyDatas:
          $ref: '#/components/schemas/CompanyData'
        addressview:
          $ref: '#/components/schemas/AddressData'
      additionalProperties: false
    DuplicateLookupRequest:
      type: object
      properties:
        applicationID:
          type: integer
          format: int32
        duplicateApplicationID:
          type: integer
          format: int32
      additionalProperties: false
    EquifaxRequest:
      type: object
      properties:
        companyID:
          type: number
          format: double
        isExisting:
          type: boolean
        isReport:
          type: boolean
      additionalProperties: false
    AddressLookupRequest:
      type: object
      properties:
        postCode:
          type: string
      additionalProperties: false
    CompanyLookupRequest:
      type: object
      properties:
        name:
          type: string
      additionalProperties: false
    ApplicationNotesDetailsRequest:
      type: object
      properties:
        applicationId:
          type: number
          format: double
        categoryId:
          type: integer
          format: int32
        noteDetails:
          type: string
        createdBy:
          type: string
      additionalProperties: false
    ApplicationsSearchRequest:
      type: object
      properties:
        pageIndex:
          type: integer
          format: int32
        itemsPage:
          type: integer
          format: int32
        surname:
          type: string
        forename:
          type: string
        postcode:
          type: string
        email:
          type: string
        town:
          type: string
        phone:
          type: string
        companyId:
          type: number
          format: double
          nullable: true
        companyName:
          type: string
        accountNo:
          type: string
        sortcode:
          type: string
        appNo:
          type: number
          format: double
          nullable: true
        loanStatus:
          type: integer
          format: int32
          nullable: true
        appStatus:
          type: integer
          format: int32
          nullable: true
        appDate:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: "JWT Authorization header using the Bearer scheme. \r\n\r\n Enter\
        \ 'Bearer' [space] and then your token in the text input below.\r\n\r\nExample:\
        \ \"Bearer 12345abcdef\""
      name: Authorization
      in: header